Refactor components to be functional and use hooks
This refactor maintains the same behavior of these components, while
changing from class components into functional components. State is
managed within these components using hooks.

The album is left as a one-way change from Beatles to Beach Boys, as
the original component did. Since the behavior was simple, I used a
basic useState hook for this component.

The counter had a slightly more complex behavior, with two different
side effects which could manipulate the state (`add` and `subtract`),
so I used the useReducer hook for this component to make the behavior
explicit, while keeping it organized.

The surprise component had a side effect which needed to be fired at
a specific point in the component lifecycle, so I pulled in the useEffect
hook. It only needs to execute once, upon unmount, so it has no dependencies
to list in the dependency array. Since there is nothing for it to do
until it is removed from the screen, the effect is merely a single
return statement containing a function to execute the alert.

For the toggle component, it was again a simple boolean state to keep
track of with a single behavior (flipping the boolean), so I leveraged
the useState hook again. I also simplified the logic a bit in the
component where there were double negatives and unnecessary empty strings.